Move SDK to pkg
